move the autogenerated file
